requirements.txt
accelerate==1.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.44.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_bleach_1737382993/work
blobfile==3.0.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
charset-normalizer==3.4.0
citepy==0.5.0
cloudpickle==3.1.0
comm==0.2.2
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1731428301407/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
dataclasses==0.6
debugpy==1.8.7
decorator==5.1.1
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dill==0.3.9
executing==2.1.0
fairscale==0.4.13
fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1733235979760/work/dist
filelock==3.16.1
fire==0.7.0
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1731643424073/work
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
imageio==2.36.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1733328101776/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266942/work
llama_models==0.0.49
llama_stack==0.0.49
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_mistune_1742402716/work
mizani @ file:///home/conda/feedstock_root/build_artifacts/mizani_1729772073073/work
MLXP==1.0.1
mpmath==1.3.0
munkres==1.1.4
-e git+ssh://git@github.com/mb208/NATS.git@530d870e579591ec984bdaa23a4a13787ae86bf6#egg=nats
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1734628800805/work
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_nbconvert-core_1738067871/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1733402752141/work
nest-asyncio==1.6.0
networkx==3.4.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1730588037451/work/dist/numpy-2.1.3-cp312-cp312-linux_x86_64.whl#sha256=1e120dee91af211b94f044893ab48bdd16fef819ebe6bad4ba71f6a6925bdb9e
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.3.3
omegaconf==2.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1731802491770/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1726878417179/work
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso==0.8.4
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1731432186484/work
pexpect==4.9.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1729065636207/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
platformdirs==4.3.6
plotnine @ file:///home/conda/feedstock_root/build_artifacts/plotnine_1732633245510/work
ply==3.11
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1728880423364/work
PySide6==6.8.0.2
pythae==0.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1743831245863/work
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1729480682811/work/dist/scipy-1.14.1-cp312-cp312-linux_x86_64.whl#sha256=5dc72baa2d67ff6e6fc7aa7dd0fe21390ed1e3400bdeddd731d9d88066ffbd26
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1714494649443/work
sentencepiece==0.2.0
setuptools==75.3.0
shiboken6==6.8.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1727986688191/work
submitit==1.5.2
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2 @ file:///home/conda/feedstock_root/build_artifacts/tinycss2_1729802851396/work
tinydb==4.8.2
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724956131631/work
tqdm==4.67.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.46.2
triton==3.2.0
typing_extensions==4.12.2
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1727140567071/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1729704561929/work
urllib3==2.2.3
wcwidth==0.2.13
webencodings @ file:///home/conda/feedstock_root/build_artifacts/webencodings_1733236011802/work
wheel==0.44.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
